Revert "Squashed commit of the following:"
This reverts commit 6e543debde9a45fded9f210d6cf3b49901b58daa.

+++ src/server/shared/SharedDefines.h
@@ -3475,7 +3475,7 @@ inline TeamId GetTeamId(PvPTeamId teamId)
 }
 
 // indexes of BattlemasterList.dbc
-enum BattlegroundTypeId : uint16
+enum BattlegroundTypeId : uint8
 {
     BATTLEGROUND_TYPE_NONE     = 0, // None
     BATTLEGROUND_AV            = 1, // Alterac Valley
